pyproject.toml
[tool.poetry]
name = "iup"
version = "0.1.0"
description = ""
authors = ["Edward Schrom <tec0@cdc.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^1.7.1"
numpy = "^2.1.1"
datetime = "^5.5"
scikit-learn = "^1.5.2"
altair = "^5.4.1"
vl-convert-python = "^1.7.0"
pyyaml = "^6.0.2"
pytest = "^8.3.3"
nisapi = {git = "https://github.com/CDCgov/nis-py-api"}

[tool.poetry.group.dev.dependencies]
detect-secrets = "^1.5.0"
pytest = "^8.3.3"
ipykernel = "^6.29.5"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
